celestial_gateway/requirements.txt
fastapi
sqlalchemy
passlib
python-jose
uvicorn